Rename files
cli.py -> spotify-dl.py
main.py -> core.py

Now they make more sense
